Add screen reader support for 'New' feature badging.
See go/newlabel for more information.

This CL adds "this is a new feature" to
the text read by the screen reader, after the hotkey but before the
"item M of N" description.

So for example, a tab context menu item might look like this:
  ...
  Add tab to Read Later [New]     CTRL+SHIFT+Z
  Move tab to new group [New]
  ...

As you scroll through, you will hear:
  "Add tab to read later, control plus shift plus Z, this is a new
   feature, item 4 of 7"
  "Move tab to new group, this is a new feature, item 5 of 7"

Note: The screenshot in the associated png file is of the visual presentation
of the badge, as the actual text read does not appear on-screen.

Bug: 1114082
Change-Id: I758d0b6cb92856e62bde87f9cd55553feb281b3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343008
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796157}