Move NavigationRequest tracing to use to use navigation id.
The NavigationThrottleRunner is tracing events related to the
NavigationRequest, however using the delegate pointer does not work to
associate the two logs together. This CL changes both NavigationRequest
and NavigationThrottleRunner to use the navigation id as the common
identifier to tie the tracing events together.

Bug: 1043616
Change-Id: I84594851724f1026fa9c9a5409705a127347a251
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343564
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796169}