weblayer: makes clearing browsing data remove favicons
This takes the heavy handed approach of deleting *all* favicon
data any time clearing data is requested. To do otherwise would
be challenging as there isn't enough information in the db to
really know which favicons should be deleted in a range. The
chrome side keys off visits to track this information, but we
don't have that.

BUG=1076463
TEST=ProfileBrowserTest.ClearBrowsingDataDeletesFavicons

Change-Id: I033ece1197889ff7f5f5edd864fc2f84e27606fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340870
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796182}