Simplified domain: fix registrable domain substring finding
Previously, we were searching for the registrable domain from the left
end of the text and using that as our simplified domain bounds. This
is incorrect in the case that the registrable domain is also a
subdomain, e.g. example.com.example.com. Now we instead search from
the right end of the hostname.

Bug: 1111902
Change-Id: I0fe97036986688b19f3a2315a32f413c8bf50b68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343572
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796191}