Removed legacy save-copy delegate function
crrev/c/2316189 removed the need for a save copy delegate function by
adding the save as code path, to make sure TOT wasn't broken it left the
old save copy delegate function as is until the next version of the
media app was rolled in. crrev/c/2344327 should update the version of
media app so it no longer uses the "save copy" function and as such its
now safe to delete it.

Change-Id: I5f7a67ffec261a8d76c1c1e6e4ff2e82851f2743
BUG: b/161193454
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2344436
Reviewed-by: Patti <patricialor@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Commit-Queue: Zain Afzal <zafzal@google.com>
Cr-Commit-Position: refs/heads/master@{#796302}