gpu: Fix error with tracking fences for the primary plane.
We were incorrectly using fences from the video buffer for initializing
the fences used for the primary plane. While this was hitting a debug
assert it didn't cause any errors in the metric since the primary plane
is always scheduled last, so the fences eventually used in the metric
did belong to the primary plane.

R=vikassoni@chromium.org

Bug: 1100442
Change-Id: I7412c4f81d6356849a1b3b42c9af94c06d565882
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2344007
Auto-Submit: Khushal <khushalsagar@chromium.org>
Commit-Queue: vikas soni <vikassoni@chromium.org>
Reviewed-by: vikas soni <vikassoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796310}