Pass CSSPropertyName to InspectorCSSAgent::FindEffectiveDeclaration
This CL is a step towards improving the suboptimal ergonimics of
the current CSSProperty API. The situation is this:

Whenever you call CSSProperty::Get(CSSPropertyID::kVariable), the
call will succeed, but the instance you get in return is an entirely
useless instance of CSSProperty known as "the static Variable
instance". This subclass, beyond being a NOTREACHED minefield, simply
returns the wrong information "silently" in some cases. For example,
CSSProperty::Get(CSSPropertyID::kVariable).GetPropertyName()
unhelpfully returns "variable", which is not even a valid
property name.

There is a (slow) ongoing effort to remove the static Variable
instance, and this CL eliminates some potential uses of it.

In InspectorCSSAgent::setEffectivePropertyValueForNode, there is
a lookup from a string to a CSSPropertyID, and then a CSSProperty::Get
using that ID. This will not work for custom properties for the reasons
already mentioned. This CL replaces that with a call to
CSSPropertyName::From instead, which *does* handle custom properties
correctly.

In InspectorAnimationAgent::CreateCSSId, PropertyHandle::GetCSSProperty
is called via the CSSTransition. The returned instance may be the
static Variable instance, so we avoid this call, and instead use
CSSPropertyName in this function.

It is possible to get a CSSProperty instance that is valid for custom
properties (see CSSPropertyRef), but since FindEffectiveDeclaration only
cares about the name of the property anyway, it's more appropriate
to use CSSPropertyName.

Bug: 980160

Change-Id: I2291b650767b17870fc6d33e1062ea127213b757
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340039
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796332}