[Autofill Assistant] Allow multiple chips in the header.
This CL allows the header to support a full carousel.

What it does:
 - Move header chip logic into native.
 - Use RecyclerView instead of hard-coded single chip view.
 - Potentially fixes an issue where the status message does not stretch
across the whole available width.
 - Chips in the header now support change animations.

There is an issue caused by a layout resize when switching between PEEK
and non-PEEK mode. This is further discussed in the linked bug. The
workaround is to delay the layout resize until the sheet state has
settled.

Bug: b/160856358
Change-Id: Id7617ed41119b85fdaa00ebaff7e856984b74b0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288567
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796376}