Access Context Auditing: Stop use of HostContentSettingsMap on shutdown
To determine whether records should be cleared on shutdown a reference
to the HostContentsSettingMap is retained for use in a TaskRunner.
This causes a race between the service shutdown in the UI thread, and
the audit database clearing session only records.

This CL changes behavior such that the content settings at the time of
shutdown are copied from the HostContentSettingsMap for local use in
the TaskRunner.

Bug: 1113376
Change-Id: I4c259a69dbc1eff9f1e85b48f05da886e5f63957
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339662
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796389}