TWA Quality enforcement on 5xx and page not available offline
This CL adds the TWA quality enforcement for http error 5xx and for
page unavailable offline.
The offline one is mostly same as 404 one, but failed when navigation
finish with errorCode -106 (internet disconnected)

Also, changes to use different error string that use for the toast and
the message send to TWA client app.
For the toast we use the localized string since it's showing in the UI.
For the message to TWA client app, since it's used in a crash message,
we do not localize it.

Bug: 1109609
Change-Id: Ic314eeb6fdfacf6e066a6e8dbb5e72334f021e63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2336996
Commit-Queue: Ella Ge <eirage@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796395}