IdleWakeups: count processes created and destroyed
This change makes IdleWakeups count the cumulative number of processes
created and destroyed with the target executable name during the
monitoring period, and display these counts in its final summary.

This enables the measurement of process turnover resulting from
operations of interest, which would otherwise not be reflected in
total process count (as a destroyed process and created process that
both occur in between polls would cancel each other out and go
undetected). Note that it is still possible to miss a process if its
lifetime is so short that it is created and destroyed entirely
between samples.

Counting is done by comparing the previous snapshot's set of process
IDs with the target executable name against the new snapshot's.

The accuracy of the counts added by this change was verified by
comparing against those collected by an ETW trace (System Activity >
Processes) over the same period. In this comparison, IdleWakeups
identified all created and destroyed Chrome processes apart from a
utility process with a lifetime of only 0.08 seconds between samples.

Change-Id: I93d674377906f6bd543da2c7b65792bb811fd91b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342251
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796426}