Reland "Allow termina to be installed via dlc"
This is a reland of d539d6835b96902894bb70a1d2e384cf1d5d0a48

The previous attempt broke a couple of browser tests which used
RunUntilIdle to wait for the uninstall to finish. This worked fine
as long as the uninstall process ran entirely on the UI thread,
but it now posts a blocking task to the thread pool. This causes
the UI thread to become idle while waiting for the response,
causing the test to continue prematurely.

This patch replaces this with explicit usage of QuitClosure.

Original change's description:
> Allow termina to be installed via dlc
>
> Also refactor the current installer code out of CrostiniManager
>
> Bug: 953544
> Change-Id: Iad0850ffba616ac3f45adff63a5e21d57d3b5e7a
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2330111
> Commit-Queue: Fergus Dall <sidereal@google.com>
> Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
> Reviewed-by: David Munro <davidmunro@google.com>
> Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
> Reviewed-by: Nicholas Verne <nverne@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#795336}

Bug: 953544
Change-Id: I1e53391eef80fada087d7bd22ee16ff01aaccf20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2341223
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Auto-Submit: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Cr-Commit-Position: refs/heads/master@{#796446}