Fix DCHECK failure in NGLineBreaker::RewindOverflow
|NGLineBreaker::RewindOverflow| assumes that |Rewind| rewinds
the current |item_index_| to the specified value. However, at
some point |Rewind| was changed to not to rewind positioned
floats.

The DCHECK was to ensure the assumption. This sometimes hits
when the line width saturates |LayoutUnit|, and it was
considered not too harmful. But crbug.com/1114324 suggested
that it may create unexpected internal states which can lead
to crashes.

This patch removes the assumption from |RewindOverflow|.

Bug: 1114324, 1020101, 1029599
Change-Id: I221e87d4ed089ae97aa21a99c2be73dc555f6d31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2344431
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796465}