Make Mac accent colors display in tabs/toolbar in incognito theme
The issue was that previously, the tabs/toolbar area used the
NativeThemeDarkAura for colors, which was not OS accent aware. This was
replaced by a forced-to-dark version of the platform's native theme. On
Mac, I added a check to maintain the old behavior that skipped the
accent colors because there was an issue with how the light/dark theme
was applied on Mac. This was since fixed, so we can remove this check
and use the accent colors in incognito. This also makes the color
behavior more consistent between regular dark mode and incognito.

Bug: 1049360
Change-Id: If9a40955391b3c6d9603c66d4eedb99afb0af752
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337575
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Commit-Queue: Joan Barnett <johnsm@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#796480}