[iOS] Add Share URL Sharing Flow
Adding a new SharingCoordinator initializer and updating
ActivityServiceCoordinator's initializers to enable triggering a URL
sharing flow for URLs that aren't necessarily associated with a WebState.

Some activities that depended on a WebState have been disabled for that
flow (e.g. Print, Find in Page). Also, this CL makes ThumbnailGenerator
optional for ChromeActivityURLSource, as no thumbnail can be generated
without a WebState.

This CL also modifies the initialization of ActivityServiceCoordinator
to use a new ActivityParams object which will contain the properties
used to configure the various sharing flows.

Bug: 1102892
Change-Id: I0b28d5fcc7bae78e669dbfac33ba252165ce86bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2319793
Commit-Queue: Sebastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796486}