[conversions] Run database sequence at TaskPriority::BEST_EFFORT
What: Re-lower conversion database sequence task priorirty.

Why: The priority was raised to USER_VISIBLE on
https://crrev.comc/2225739 in an attempt to resolve test timeouts
seen with the feature enabled. This change did not fix the flakes, and
the flakes were later addressed separately.

This returns the priority to the intended value, as conversion
operations work fine with delays.

Bug: 1080764
Change-Id: I5f5907bd5e7711b0a6ed1db45db9f3713acf4765
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340714
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796510}