DCHECK in AccessibilityPerformAction with IA2 setSelection method
This is the corresponding IA2 fix for the selection-iframe crash here:
https://chromium-review.googlesource.com/c/chromium/src/+/2300266

The approach is different for IA2. Since IA2's setSelectionRanges
requires actual nodes to be passed in, it doesn't make sense to move
the selection range to different nodes than specified. Instead, this
change simply returns E_INVALIDARG when two nodes are provided that
exist in different trees. I generally don't like this approach, as it
is reporting implementation limitations as a user-error, but I think
it's the best approach for this particular scenario.

A browser test was added that would crash without this fix in place.

AX-Relnotes: n/a

Bug: 1110522
Change-Id: Ifa8e9e7b43c52d611d0c77b134f5cb0ab3b86357
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340291
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796515}