[ozone/wayland] Use CommitOverlays to send a frame from gpu to browser.
CommitOverlays is a more generic IPC that packages multiple overlay
configurations for a frame before sending it to browser process. This
replaces the CommitBuffer IPC.

This CL merely adds CommitOverlays. It does not send multiple buffers
using it.

This is 3/? CL for overlay forwarding using wl_subsurface.

Bug: 1063865
Change-Id: Id71d055bdb14b3d2e97cb2eef8811dca56323899
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2314665
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Maksim Sisov (GMT+3) <msisov@igalia.com>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Kramer Ge <fangzhoug@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796542}