Adds a chrome://flags entry for middle-slot promo browser commands
Adds chrome://flags/#promo-browser-commands in order to facilitate
testing the promo browser commands.

For every new command, an option should be added to this flag. Selecting
that option would replace the middle-slot promo on the NTP with a mocked
promo and replaces the promo URL with command:<ID>.

Bug: 1106421
Change-Id: I53e07d8fcdea057f43d9ff086c49a93565d79969
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346834
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796543}