Stop browser_tests leaving an open Explorer window
This change stops browser_tests from leaving a file explorer window
open after it completes.

Currently, browser test `DownloadExtensionTest.DownloadsShowFunction`
tests the "Show in Folder" functionality by opening a file explorer
window, but has no way to close this window after it completes. This
results in a stray file-explorer window. At times, the stray window
may also come with an error popup if the temp directory it points to
has already been deleted.

With this change, the call to open the file explorer is skipped if
shell operations have been disabled for testing. The call is skipped
as late as possible, in order to exercise as much of the code as
possible.

Change-Id: If015143d98a3e7e5d4a6bce958780fce6fa63f32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339599
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/master@{#796626}