Share PDFiumEngine creation between PDF plugins
Consolidates code for creating (and destroying) the main PDFiumEngine in
a common base class, PdfViewPluginBase. This base class can be used for
any code which is not specific to the Pepper (OutOfProcessInstance) or
Pepper-free (PdfViewWebPlugin) plugin implementations.

Note that this approach involves multiple inheritance of implementation,
but I believe this is a reasonable trade-off during the migration.

Bug: 1085334
Change-Id: I5b17141972a371e70cf0e6801d7dad2cb0d20aae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343573
Commit-Queue: K. Moon <kmoon@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796648}