Skip setting colorspace for YUV images
YUV textures from OES_EGL_image_external are sampled as RGB.
Technically, the driver is supposed to preserve the colorspace, so we
could still pass the primaries and transfer function. However, we don't
actually pass the colorspace to the driver, and it's unclear what
drivers would actually do if we did. So in effect, the colorspace is
undefined.

Bug: 1102317
Change-Id: I7f8b430e330fc25e3c2414ae66d1d22708e67f2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2331769
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796675}