chrome://media-app: hide delete option for files from openSavePicker().
Currently the button is available, but it does not work if the new file
is in a different folder. Clicking it generates a crash report.

Hide it by removing the functions that the UI uses to determine whether
these operations are available for a given file, after a successful
saveAs.

Bug: b/163277173
Change-Id: Iee4e6318012a314da2273bcf1a1b04414f786f4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2344272
Reviewed-by: Rachel Carpenter <carpenterr@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796699}