appcache: don't create service when its feature is disabled
Currently, disabling AppCache by turning off blink::features::kAppCache
is a shallow disabling, in that the web api and the loader no longer use
it, and its preference is disabled, but the service is still created.

This is a step towards adding data deletion behavior when this feature
is disabled.  It doesn't make sense to delete the data folder when there
is still an appcache service that exists at runtime.

This patch makes the AppCache service optional and does not create it
when the kAppCache feature is disabled.

Bug: 1081897
Change-Id: I3a2a6cf4dd5bb8654a3c0da9ce310acff91766eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339920
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Auto-Submit: enne <enne@chromium.org>
Commit-Queue: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796715}