Do more clean up in some Print Preview browser tests.
Normally, when Print Preview finishes, the browser will call
PrintPreviewDone(), which in turn calls OnClosePrintPreviewDialog().
Change PrintPreviewDialogControllerBrowserTest and
PrintRenderFrameHelperPreviewTest to do this as well.

This is in preparation for a follow-up CL that changes when webpages are
(un)paused for printing. With that CL, tests can no longer start a print
preview and leave it in an unfinished state like it does now.

Bug: 956832
Change-Id: I70908f0b86bfa288c257a65046c021cdd6138c2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340515
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796721}