Fix Meet flickering on macOS
As background, YUV 420 biplanar videos on macOS (basically all videos)
cannot be sampled directly in OpenGL. When we want to sample them, we do
a copy and YUV to RGB conversion.

In the compositor, we will draw either using OpenGL or using overlays.

We bracket our use with calls to Begin/EndSharedImageAccess using the
argument READ (for GL) or OVERLAY (for overlay). In the case of overlay
access, we skip the copy.

The DisplayResourceProvider will agglomerates the calls to
Begin/EndSharedImageAccess into a single pair -- so if there is already
a READ access active, and another READ access comes in, then we skip
the second call to Begin/EndSharedImageAccess.

The bug is that if we do OVERLAY access, and then want READ access, the
DisplayResourceProvider will say "oh, you already have access, I'll
skip the BeginSharedImageAccess call". Which means no copy, and
random texture contents.

To fix this, detect the case of transitioning from OVERLAY to READ
access, and do a EndSharedImageAccess followed by a new
BeginSharedImageAccess.

This is all fairly horrible. Once we move everything to the
SkiaRenderer, then we can remove the OVERLAY GL parameter and
forget that this episode ever existed.

Bug: 1113925
Change-Id: I25071324eb49c5fd3026a77c20034c9f5e647328
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346843
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796730}