Use immediate callback runner in testing profile
Currently |TestingProfile| initializes Chrome OS Account Manager with
|chromeos::DelayNetworkCall|, which uses |NetworkHandler|.
It leads to crashes in tests which do not initialize |NetworkHandler|
properly. (This code path was not exercised before, but in
https://crrev.com/c/2232324 we start using Account Manager in tests
and some tests crash because of this issue.)

Fix this by not using |chromeos::DelayNetworkCall| and instead supplying
a fake runner which immediately executes its callbacks.

Bug: 1068240
Change-Id: I220ebc32fa069876f4cd607ba1bae89a35ac7d41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346350
Reviewed-by: Kush Sinha <sinhak@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796760}