[PwdCheckAndroid] Check is started automatically and on restart
With this CL, the check is automatically started if a user clicks the
restart button or navigates to the password check page from a
non-safety check origin.

The check is run to completion unless all observers unsubscribe before
the end of the check (e.g. because their surface is closed).

Promoted bugs:
 * List receives full update for each new credential
 * The "last run" date range does not seem sensible

Bug: 1109691, 1092444
Change-Id: Icfb25b91508d36b938e56cea2516922089963de2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346387
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796777}