Report Copying time using UMA statistics
We want to improve the installation time for force installed
extensions. For this, we need to first gather the statistics regarding
which installation stages are time consuming. We collect the UMA
statistics for the time taken to copy the extension archive to the
working directory.

Following histogram is added:
Extensions.ForceInstalledTime.CopyingStartTo.UnpackingStart

Bug: 1102369
Change-Id: I09f27e2f656581f318f32cbe137b0d1da9ab5e62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2338794
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Cr-Commit-Position: refs/heads/master@{#796800}