[PwdCheckAndroid] Set status and initial run by referrer
This CL ensures that the referrer is considered in the decision to
run the password check.
When coming from safety check:  use last status, don't run check
When coming from anywhere else: run the check when possible.

The initial status is not relevant for the decision to run.
If it's RUNNING, the call is a noop.
If it's IDLE or cancelled, it makes sense to run a new check.
If it's any error state, rerunning might show the error was
fixed since the last attempt. Reaching a new failure won't take long.

Bug: 1109691, 1092444
Change-Id: I573a5f9f0a8861e598541fa38faa2d8bd42adba3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2348773
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796854}