Stop showing move bubble if non-opted-in user refused it too many times
This CL adds a field trial parameter to the passwords account storage
feature. It is used to limit the number of times Chrome offers a user
that hasn't gone through the opt-in flow yet to move a password to their
account. The methods from crrev.com/c/2340966 are used to query the
number of times the user has already refused the move.

No behavior is actually changed however, since the refusal counter is
not yet incremented (future CLs).

TBR=vasilii@chromium.org

Bug: 1082152
Change-Id: I248f0964e1b406026016288d5e41fb64b2a6e10d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2345309
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Victor Vianna <victorvianna@google.com>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796871}