Uses the same definition of an embedded object in both implementations of AXPosition
This is the first step in the process of merging BrowserAccessibilityPosition and AXNodePosition into one class.
We need to adopt the same definition of an embedded object in both places.
This is the definition used by IA2 and ATK hypertext, which makes all non-textual nodes that are exposed to the platform APIs
into embedded objects.
Additionally, we include the definition used by UIA, which makes all empty leaf objects (objects with ignored descendants)
as embedded objects.

AX-Relnotes: n/a.

R=dmazzoni@chromium.org, aleventhal@chromium.org

Change-Id: I19bbc26f40d37a07b7dd6436fb0025505a0cfd89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346653
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796893}