[Paint Preview] First Meaningful Paint Swap Trigger
This is an alternative/complimentary change to
http://crrev.com/c/2346905

Using First Meaningful Paint has two benefits.

1. It waits for key content to paint.
2. It doesn't wait for total loading to stop just down to a couple
   network requests. This is good in that we stop showing the preview
   earlier (which is desirable if the page is actually painted sooner).

I think both this and LoadingProgress are feasible. This one is more
promising in the sense that it takes a guess as to when the painting
is mostly done whereas the PageLoadProgress variant is mostly about
network requests being done.

One option might be to combine the two and take the later of the
signals? With these changes it might be worth making the removal
delay 0 ms or at least smaller.

Bug: 1110360
Change-Id: I6c4400166d0bfef57160424de2238bcbe238949a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346863
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796953}