Pointer Events tiltY is reversed on Mac
On Mac, tiltY is positive when the pen is tilted away from
the user. Based on Pointer Event spec, tiltY should be positive when
the pen is tilted towards the user.

On Mac, by default Y increases going up, while in Chromium Y increases
going down. NSView has isFlipped property which controls if Y increases
going up (false) or going down (true).

In order to fix the bug, we check NSView's isFlipped flag and if it is
false we reverse the sign of tiltY to accommodate the difference in the
direction of Y.

Bug: 1111347
Change-Id: I108ba496adc6f951277e1c9ddb7eaf6483e522f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2348544
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796963}