Disable opening extensionless downloads on Windows
This change makes the "Open" option in the downloads bar act like
"Show in Folder" for files without file extensions on Windows.

This fixes an issue where Chrome could unexpectedly open the wrong
file when attempting to open an extensionless file while an executable
of the same name was present in the Downloads folder. For example,
downloading and opening "somefile" while "somefile.exe" was already
present in Downloads would result in "somefile.exe" being run. This
behavior is due to Windows' shell open functionality, which appears to
have special logic to interpret extensionless commands as potential
executables.

Prior to this change, if an extensionless file was opened without an
executable of the same name present (the more common scenario),
Windows would display error "No application is associated with the
specified file for this operation", so this change does not remove any
functionality apart from the ability to trigger this error popup.

Showing the extensionless file in folder sets the user up nicely to
open the file however they want from Windows Explorer via right-click
> Open With.

Bug: 1092518
Change-Id: I1a3bb519867548c916070cbfebd64d0017f5937b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2321528
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796968}