Get LayerTreeHostImpl out of the input business
This CL begins the process of splitting input handling out of
LayerTreeHostImpl. To keep things simple and easy to review, this CL is
a mostly mechanical disentangling of input code into a new class,
ThreadedInputHandler. The class still lives in layer_tree_host_impl.cc
and is still tightly coupled with it. Future CLs will move this into a
separate file and put up proper interfaces between compositing and
input. For now there is no behavioral changes, the two classes delegate
calls to each other as appropriate.

The ThreadedInputHandler has methods shadowing the InputHandler
interface, and LayerTreeHostImpl now delegates to it, but
ThreadedInputHandler doesn't yet implement the interface. A future CL
will make it implement InputHandler and pass it around instead of LTHI.

The end state is to have ThreadedInputHandler be a standalone object
used by InputHandlerProxy to affect input-related changes. This split
deliberately leaves anything graphics or animation related in LTHI.

Bug: 915926
Change-Id: Iabc08c03fb618b1f7e9406b5ab7fa7ab64d60913
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2347371
Reviewed-by: Xida Chen <xidachen@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796970}