[omnibox] Send oft=2 for on-clobber ZeroSuggest requests
This is a trivial CL that just sets oft=2 for on-clobber ZeroSugest
requests.

I had originally thought that maybe we would want to add a feature flag
that makes it still send oft=1, but now I think that would just be
confusing, and it would be better to quickly update the server-side to
properly handle both oft=1 and oft=2. As a first pass, they could just
be treated the same.

Bug: 1106096
Change-Id: Idcbfc55e164865d10a000b62a585898bbf9c20d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2350308
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797060}