[iOS] Initialize SafeBrowsingServiceImpl's SharedURLLoaderFactory eariler
This CL initializes SafeBrowsingServiceImpl's
SharedURLLoaderFactory in Initialize(), rather than two
thread hops after Initialize(). This ensures that
GetURLLoaderFactory() has a non-null return value when called
shortly after Initialize().

This fixes crashes that occur when CreateUrlChecker() is called
soon after Initialize(), resulting in the creation of a
RealTimeUrlLookupService that has a null URLLoaderFactory and
crashes when it tries to use it.

Bug: 1114706
Change-Id: I9142c82979954a272f351c14db20246730f720db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346862
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797066}