Remove default apps checking, and convert all apps to the standard icons
Since we have the circle checking, we can call the CreateStandardIcon
interface to convert all apps (including the essential apps/default
apps) to the standard icons, and the CreateStandardIcon interface can
check whether the icon is circle or not to apply shrink or not. So the
default apps checking can be removed.

BUG=1114771

Change-Id: I6c60fa8127c6f22e19824e21bffc8720100576f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2351259
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797138}