[Autofill Assistant] Remove intent validation for password scripts
As password scripts have a confirmation step, we don't need to validate
an intent to change a password.

This CL removes timer that will rest password state. Password manager
will be suspended for all autobot flows.

This CL also changes what origin is used for password store request. If
a password script is started from Settings, GetLastCommittedURL is not
set yet. So, we have to use the deep url provided.

Bug: 1086109
Change-Id: I636d3f7891d98212a3998c450b94f72a0fc51573
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2348771
Commit-Queue: Milica Selakovic <selakovic@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797168}