Fix FontPlatformDataMacTest.VariableOpticalSizingThreshold for Big Sur
On Big Sur, the system font has a full, real, 'opsz' axis rather than
the 10.15 trick of gluing together two optical sizes. Adjust
the test to match.

Do some cleanup in the main .mm file. Switch from VLAs in
VariableAxisChangeEffective() to WTF::Vector (as is used in the
rest of the file) as VLAs are not allowed in Chromium nor Blink,
and use #include as the style is to use #import for Objective-C
headers.

Bug: 1101448
Change-Id: I6043b53b9ac07cdddd082a19abf7bc8cbe3af58b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2351299
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Auto-Submit: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797228}