Record "move password to account" bubble offered to non opted-in user
The CL calls the incrementer from crrev.com/c/2340966 (later renamed).
Now, a per-Gaia-account counter is incremented whenever the move bubble
is shown and the user is not opted-in to the account-scoped passwords
storage. The bubble won't be shown for users whose count is above a
certain threshold (see crrev.com/c/2345309).

Bug: 1082152
Change-Id: Id0ef7fd6b85598a1cb0651435a5c6f9bf6214049
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2351995
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#797268}