[Nearby] Add certificate methods to hash authentication token
The Nearby Share protocol uses HKDF with an authentication token, using
a certificate's secret key as the salt and a trivial info, to generate a
hash. Both parties--the receiver and sender--confirm this hash. This,
along with a standard signing/verification flow, is used to establish
an authenticated connection between devices. See go/nearby-chrome-cert.

This CL implements a hashing method for both the private and
decrypted-public certificate classes. We add tests using test vectors
that are also used in GmsCore unit tests (cl/326170193) to ensure
consistency across platforms.

Fixed: 1115320
Change-Id: I5aa9f7de5340a537a57c323a5afb97f6a83a4b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2351709
Reviewed-by: Alex Chau <alexchau@chromium.org>
Commit-Queue: Josh Nohle <nohle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797270}