macviews: remove NativeWidget damage rect propagation tests
These tests were asserting that, when Views damages a specific rect
within the BridgedContentView, AppKit was propagating the damaged
rect all the way back to [BridgedContentView drawRect:]. There are
two issues with this:

1) BridgedContentView does not implement drawRect:, and
2) Since BridgedContentView is layer-hosted AppKit is not supposed
   to track damage to it and is *supposed* to always call drawRect:
   on the entire thing

It appears that AppKit in <= 10.15 did bother to track the damage
and invoke drawRect: only on the damaged area even for layer-hosted
views, presumably as an optimization for misguided code that was
improperly using layers. As of 11 AppKit does not do this, so the test
(which was asserting behavior that was both incorrect and irrelevant
to us) is broken.

This change deletes these tests.

Bug: 1114240
Change-Id: I22ede812cdf98707cc22bca1238f59856a371d36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352431
Commit-Queue: Robert Liao <robliao@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797329}