Transfer ownership of //third_party/grpc
The Chromoting team was the owner of //third_party/grpc, as we first
introduced the gRPC dependency in Chromium. However, we have recently
migrated our app from using gRPC to just using plain Protobuf requests
over HTTP due to some network security concerns and the burden of
maintaining the dependency in Chromium third_party.

It appears that chromecast and libassistant is currently the only other
users of //third_party/grpc. As previously discussed in the email and
code review comments, this CL transfers the ownership of
//third_party/grpc to chromecast and libassistant.

Bug: 1103416
Change-Id: I07ae66ab05b7fc57d3b4fc3e6222585dacbad530
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2350189
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Yue Li <updowndota@chromium.org>
Reviewed-by: Luke Halliwell (slow) <halliwell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797360}