cr_components/chromeos: Use internet-config-dialog host WebUI for tests.
Networking components have a lot of dependencies that need to be in
place for testing, so use internet-config-dialog as the host instead of
OS Settings.

Note that ideally this would not be required for testing, but with
Polymer2 going away soon, it's not worth building more infra to support
this case.

R=khorimoto@chromium.org

Bug: none
Test: none
Change-Id: I06ca337eb0c18e362d91d8ddee81ffe2707dd0b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352551
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: James Hawkins <jhawkins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797382}