Fix the crash caused on shutdown by OnLogChange method
When the browser is shutdown, we start destroying the objects. In
ExtensionInstallEventLogManger, we used to destroy |logger_| after the
|extension_log_upload_| pointer. This could cause to a crash because
we try add a LOGOUT event before the shutdown. And while adding the
LOGOUT event, ExtensionInstallEventLogManager::Add method is invoked
which posts a task to invoke
InstallEventLogManagerBase::LogUpload::OnLogChange. Thus we should
destroy the |extension_log_upload_| pointer after destroying the
|logger_| pointer so that we add the LOGOUT event and prevent the
crash. Similarly, this should be fixed for ARC++ reporting.

Bug: 1114191
Change-Id: I6e88f5747b6a5b7a36e1dcee3083204833c9aeff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346306
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Anqing Zhao <anqing@chromium.org>
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Cr-Commit-Position: refs/heads/master@{#797388}