Protect KeyChainCertSelectionCallback against double calls.
This is a speculative fix for https://crbug.com/1114334. Logs attached
to crash reports indicate that the KeyChainAliasCallback is being
invoked twice. Specifically, "No client certificate selected" appears
twice in the log, and in one instance we see them appear within 1ms of
each other.

It's possible that Chrome's maybeShowCertSelection is accidentally
invoking the callback one of those times, or that Android's KeyChain
framework is double-invoking the callback. The docs do not specify how
many times the callback may be invoked.

This CL also switches logging to org.chromium.base.Log and limits the
TAG to 20 characters, as required by presubmits.

Bug: 1114334
Change-Id: I3e5e1f318f579b0b7a4521f058ba1f0a7224c487
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2344624
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797414}