New exit fullscreen message for CrOS
See attached bug for the problem statement and some of the rationale for
this change.

Handles a case where we are not going into immersive mode, but also
going fullscreen in a way that the default key to restore the browser is
not ESC (as it is for media, etc.) but F11/Fullscreen.

In this edge case, on ChromeOS, the message - "Press [F11] to exit full
screen" - is displayed and announced to the screen reader. The message
uses the default accelerator (which is still F11 on ChromeOS). This
could be confusing to users on Chromebooks where there are no physical
function keys but there are media keys, one of which is the fullscreen
key.

We cannot just change the default accelerator to Fullscreen however
since the media key is translated to a function key at a lower level;
there is no "fullscreen" key as far as Chrome is concerned. To handle
this, on ChromeOS only, we substitute the name of the key -
"Fullscreen" - for the name of the accelerator.

Bug: 1110468
Change-Id: Id0764f3eb68c8dc49baf5cbe156cbd22e66362d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2347962
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797457}